add check to see if existing processes exist

+++ deploy.sh
@@ -143,7 +143,10 @@ function create_launch_script () {
     cd $VM_PROJECT_PATH
     source $VM_PROJECT_PATH/.env
     source $VM_HOME_DIR/venv/bin/activate
-    sudo kill ${gunicorn_pid}
+    if [ ! -z ${gunicorn_pid} ]; then
+        printf "\nKilling previous instances...\n"
+        sudo kill ${gunicorn_pid}
+    fi
     sudo $VM_HOME_DIR/venv/bin/gunicorn -b 0.0.0.0:$DEPLOYMENT_PORT --env APP_CONFIG=${DEPLOYMENT_ENV} --daemon app:APP
     printf "\n\n***************************************************\n\t\tDeployment Succeeded.\n***************************************************\n\n"
 EOF
